+++ addon/content/components/contactDetail.jsx
@@ -5,9 +5,9 @@
 import React from "react";
 import * as ReactRedux from "react-redux";
 import PropTypes from "prop-types";
-import { summaryActions } from "./reducer-summary.js";
+import { summaryActions } from "../reducer/reducer-summary.js";
 import { SvgIcon } from "./svgIcon.jsx";
-import { browser } from "./es-modules/thunderbird-compat.js";
+import { browser } from "../es-modules/thunderbird-compat.js";
 
 function _ContactDetail({
   name,
