+++ addon/content/components/conversation/conversationFooter.jsx
@@ -5,8 +5,8 @@
 import React from "react";
 import * as ReactRedux from "react-redux";
 import PropTypes from "prop-types";
-import { summaryActions } from "./reducer-summary.js";
-import { browser } from "./es-modules/thunderbird-compat.js";
+import { summaryActions } from "../../reducer/reducer-summary.js";
+import { browser } from "../../es-modules/thunderbird-compat.js";
 
 class _ConversationFooter extends React.PureComponent {
   constructor(props) {
