+++ addon/content/components/conversation/conversationWrapper.jsx
@@ -7,9 +7,9 @@ import * as ReactRedux from "react-redux";
 import PropTypes from "prop-types";
 import { ConversationFooter } from "./conversationFooter.jsx";
 import { ConversationHeader } from "./conversationHeader.jsx";
-import { messageActions } from "./reducer-messages.js";
-import { MessageList } from "./messageList.jsx";
-import { browser } from "./es-modules/thunderbird-compat.js";
+import { messageActions } from "../../reducer/reducer-messages.js";
+import { MessageList } from "../message/messageList.jsx";
+import { browser } from "../../es-modules/thunderbird-compat.js";
 
 class _ConversationWrapper extends React.PureComponent {
   constructor(props) {
