+++ addon/content/components/message/messageActionButton.jsx
@@ -4,8 +4,8 @@
 
 import React from "react";
 import PropTypes from "prop-types";
-import { SvgIcon } from "./svgIcon.jsx";
-import { browser } from "./es-modules/thunderbird-compat.js";
+import { SvgIcon } from "../svgIcon.jsx";
+import { browser } from "../../es-modules/thunderbird-compat.js";
 
 const ActionsToInfoMap = {
   draft: {
