+++ addon/content/components/message/messageDetails.jsx
