+++ addon/content/components/message/messageHeader.jsx
@@ -5,12 +5,12 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import PropTypes from "prop-types";
-import { ContactDetail } from "./contactDetail.jsx";
-import { messageActions } from "./reducer-messages.js";
+import { ContactDetail } from "../contactDetail.jsx";
+import { messageActions } from "../../reducer/reducer-messages.js";
 import { MessageHeaderOptions } from "./messageHeaderOptions.jsx";
 import { MessageTags, SpecialMessageTags } from "./messageTags.jsx";
-import { SvgIcon } from "./svgIcon.jsx";
-import { browser } from "./es-modules/thunderbird-compat.js";
+import { SvgIcon } from "../svgIcon.jsx";
+import { browser } from "../../es-modules/thunderbird-compat.js";
 
 /**
  * Normalize a contact into a string (used for i18n formatting).
