+++ addon/content/components/quickreply/quickReply.jsx
