+++ addon/content/reducer/reducer-messages.js
@@ -6,7 +6,7 @@
 
 import { summaryActions } from "./reducer-summary.js";
 import * as RTK from "@reduxjs/toolkit";
-import { browser as _browser } from "./es-modules/thunderbird-compat.js";
+import { browser as _browser } from "../es-modules/thunderbird-compat.js";
 
 // Prefer the global browser object to the imported one.
 window.browser = window.browser || _browser;
